Add an option to run golangci-lint from the directory of config file
golangci-lint has issues running in subdirectories of the project,
e.g. golangci/golangci-lint#3656, golangci/golangci-lint#3717, so
running it from the directory where configuration file is located
is the only option for some projects.

Make it work as an option, because not every project triggers this
problem in golangci-lint, so for these projects overhead of
looking up directory of config file and adjusting the paths after
is not needed.