feat: Support projects with multiple go.mod files
golangci-lint always searches the cwd upwards for the closest go.mod and
.golangci.yaml file. This means that when linting a .go file, we can
safely run golangci-lint from the directory in which the file is
contained. In projects with a single go.mod in its root, golangci-lint
will find it. In monorepos with a go.mod closer to the file under lint,
golangci-lint will find it as well.

The rootDir prefix has to be stripped from the file, since otherwise it
will not match the filename reported by golangci-lint.